Move greedy related tests into index.js
